feature aggregation

+++ sent_distribution_validating.py
@@ -439,7 +439,7 @@ else:
       f"EleutherAI/pythia-{args.model_size}-deduped",
       revision="step143000",
       cache_dir=f"./pythia-{args.model_size}-deduped/step143000",
-    ).cuda(args.cuda).half().eval()
+    ).cuda(args.cuda).eval()
     model = model.to_bettertransformer()
 
     tokenizer = AutoTokenizer.from_pretrained(
